# docs/api/openapi.yaml
openapi: "3.1.0"
info:
  title: TrustInbox Provider API
  version: "1.0.0"
  description: |
    REST API for verified service providers to send notifications, manage
    callback requests, bots, webhooks, campaigns, and view analytics.
    All endpoints require a valid API key via the `X-API-Key` header.
  contact:
    name: TrustInbox Engineering
  license:
    name: Proprietary
servers:
  - url: http://localhost:8080
    description: Local development
security:
  - apiKey: []
components:
  securitySchemes:
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: Service provider API key in the form `{prefix}.{secret}`
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
      required: [error]

    # ── Notification ─────────────────────────────────────────
    CreateNotificationRequest:
      type: object
      required: [userId, category, title, body]
      properties:
        userId:
          type: string
          format: uuid
        category:
          type: string
          enum: [PERSONAL, SERVICE_PROVIDER, ADVERTISEMENT]
        title:
          type: string
        body:
          type: string
        priority:
          type: string
          enum: [LOW, NORMAL, HIGH, URGENT]
        metadata:
          type: object
          additionalProperties:
            type: string
    CreateNotificationResponse:
      type: object
      properties:
        notificationId:
          type: string
        status:
          type: string
        rejectionReason:
          type: string
...
  parameters:
...
    offsetParam:
      name: offset
      in: query
      schema:
        type: integer
        default: 0
...
paths:
...
  /api/v1/notifications:
    post:
      summary: Send a notification to a user
      operationId: createNotification
      tags: [Notifications]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateNotificationRequest"
      responses:
        "201":
          description: Notification created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateNotificationResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Invalid API Key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    get:
      summary: List notifications
      operationId: listNotifications
      tags: [Notifications]
      parameters:
        - name: userId
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: category
          in: query
          schema:
            type: string
        - $ref: "#/components/parameters/limitParam"
        - $ref: "#/components/parameters/offsetParam"
      responses:
        "200":
          description: Notification list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationList"

  # ── Callbacks ────────────────────────────────────────────
  /api/v1/callbacks:
    post:
      summary: Create a callback request for a user
      operationId: createCallbackRequest
      tags: [Callbacks]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCallbackRequest"
      responses:
        "201":
          description: Callback request created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateCallbackResponse"
    get:
      summary: List callback requests
      operationId: listCallbackRequests
      tags: [Callbacks]
      parameters:
        - name: userId
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
            enum: [PENDING, APPROVED, REJECTED, EXPIRED]
        - $ref: "#/components/parameters/limitParam"
        - $ref: "#/components/parameters/offsetParam"
      responses:
        "200":
          description: Callback request list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CallbackRequestList"
  /api/v1/callbacks/{id}:
    get:
      summary: Get a specific callback request
      operationId: getCallbackRequest
      tags: [Callbacks]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Callback request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CallbackRequest"
        "404":
          description: Not found

  # ── Messages ─────────────────────────────────────────────
  /api/v1/messages:
    post:
      summary: Send a message in a conversation
      operationId: sendMessage
      tags: [Messages]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMessageRequest"
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
    get:
      summary: List messages in a conversation
      operationId: listMessages
      tags: [Messages]
      parameters:
        - name: conversationId
          in: query
          required: true
          schema:
            type: string
        - $ref: "#/components/parameters/limitParam"
        - $ref: "#/components/parameters/offsetParam"
      responses:
        "200":
          description: Message list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageList"

  # ── Documents ────────────────────────────────────────────
  /api/v1/documents:
    post:
      summary: Share a document with a user
      operationId: shareDocument
      tags: [Documents]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ShareDocumentRequest"
      responses:
        "201":
          description: Document shared
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShareDocumentResponse"

  # ── Campaigns ────────────────────────────────────────────
...
  /api/v1/campaigns/{id}/analytics:
    get:
      summary: Get analytics for a specific campaign
      operationId: getCampaignAnalytics
      tags: [Campaigns]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Campaign analytics
          content:
            application/json:
              schema:
                type: object
                properties:
                  totalTargets:
                    type: integer
                  totalSent:
                    type: integer
                  totalDelivered:
                    type: integer
                  totalRead:
                    type: integer
                  deliveryRate:
                    type: number

  # ── Webhooks ─────────────────────────────────────────────
  /api/v1/webhooks:
    post:
      summary: Create a webhook subscription
      operationId: createWebhookSubscription
      tags: [Webhooks]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWebhookRequest"
      responses:
        "201":
          description: Subscription created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookSubscription"
    get:
      summary: List webhook subscriptions
      operationId: listWebhookSubscriptions
      tags: [Webhooks]
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - $ref: "#/components/parameters/limitParam"
        - $ref: "#/components/parameters/offsetParam"
      responses:
        "200":
          description: Subscription list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookSubscriptionList"
...
  /api/v1/bots:
    post:
      summary: Create a bot
      operationId: createBot
      tags: [Bots]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBotRequest"
      responses:
        "201":
          description: Bot created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bot"
    get:
      summary: List bots
      operationId: listBots
      tags: [Bots]
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [DRAFT, ACTIVE, PAUSED, ARCHIVED]
        - $ref: "#/components/parameters/limitParam"
        - $ref: "#/components/parameters/offsetParam"
      responses:
        "200":
          description: Bot list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BotList"
  /api/v1/bots/{id}:
    get:
      summary: Get a bot
      operationId: getBot
      tags: [Bots]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Bot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bot"
    put:
      summary: Update a bot
      operationId: updateBot
      tags: [Bots]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateBotRequest"
      responses:
        "200":
          description: Updated bot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bot"
    delete:
      summary: Delete a bot
      operationId: deleteBot
      tags: [Bots]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
  /api/v1/bots/{id}/config:
    get:
      summary: Get bot configuration
      operationId: getBotConfiguration
      tags: [Bots]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Bot configuration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BotConfiguration"
    put:
      summary: Update bot configuration
      operationId: updateBotConfiguration
      tags: [Bots]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BotConfiguration"
      responses:
        "200":
          description: Updated configuration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BotConfiguration"
...
  /api/v1/bots/{id}/analytics:
    get:
      summary: Get bot analytics
      operationId: getBotAnalytics
      tags: [Bots]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Bot analytics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BotAnalytics"

  # ── Analytics ────────────────────────────────────────────
...